+++ layrzsdk/entities/charts/bar.py
@@ -57,22 +57,14 @@ class BarChart:
   def render(self):
     """
     Render chart to a Javascript Library.
-
-    With less than 10.000 points (in X Axis), will return ApexCharts configuration. Else will return Google Charts
+    Currently only available for ApexCharts.
     """
-
-    if len(self.y_axis) >= 1:
-      return {
-        'library': 'APEXCHARTS',
-        'configuration': self.__render_apexcharts(len(self.y_axis[0].data) * len(self.y_axis) > 9_000)
-      }
-
     return {
       'library': 'APEXCHARTS',
       'configuration': self.__render_apexcharts()
     }
 
-  def __render_apexcharts(self, large_dataset=False):
+  def __render_apexcharts(self):
     """
     Converts the configuration of the chart to Javascript library ApexCharts.
     """